Remove content app and move to core
