+++ .secrets.baseline
@@ -151,7 +151,7 @@
         "filename": "core/settings.py",
         "hashed_secret": "1ee34e26aeaf89c64ecc2c85efe6a961b75a50e9",
         "is_verified": false,
-        "line_number": 210
+        "line_number": 209
       }
     ],
     "docker-compose.yml": [
@@ -207,5 +207,5 @@
       }
     ]
   },
-  "generated_at": "2024-11-18T10:26:49Z"
+  "generated_at": "2024-11-19T08:59:11Z"
 }
