--- content/context_processors.py
@@ -1,5 +1,0 @@
-from . import constants
-
-
-def content(request):
-    return vars(constants)
