+++ core/context_processors.py
@@ -1,5 +1,7 @@
 from django.conf import settings
 
+from . import constants
+
 
 def environment(request):
     return {
@@ -7,3 +9,7 @@ def environment(request):
         "DEBUG": settings.DEBUG,
         "is_embedded": True,
     }
+
+
+def content(request):
+    return vars(constants)
