+++ core/settings.py
@@ -53,7 +53,6 @@ INSTALLED_APPS = [
     "dsfr",
     "django.forms",
     "colorfield",
-    "content",
     "core",
     "qfdmd",
     "qfdmo",
@@ -165,7 +164,7 @@ def context_processors():
         "django.contrib.auth.context_processors.auth",
         "django.contrib.messages.context_processors.messages",
         "core.context_processors.environment",
-        "content.context_processors.content",
+        "core.context_processors.content",
         "dsfr.context_processors.site_config",
     ]
